🔥 Chore: Done adding feature map to request
