+++ client/pages/requests/[id].js
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import { useContext, useState, useRef } from 'react';
+import { useContext } from 'react';
 import { useQuery, useMutation } from 'react-query';
 import { useRouter } from 'next/router';
 import { Send, File, Check, MessageSquare } from 'react-feather';
@@ -12,10 +12,10 @@ import ErrorMessage from '../../components/ui/ErrorMessage';
 import NavBar from '../../components/NavBar';
 import Message from '../../components/ui/Message';
 import Header from '../../components/Header';
-import CommentList from '../../components/Messaging/Comments';
 import Map from '../../components/Map';
 import useViewport from '../../hooks/useViewport';
 import Mark from '../../components/Map/Mark';
+import Comments from '../../components/Request/Comments';
 
 export const getServerSideProps = async (ctx) => {
   // TODO: Get the data from the server here using ctx.params.id
@@ -39,46 +39,10 @@ function RequestId({ id }) {
   const { isLoading, isError, data } = useQuery('request', () =>
     useAxios({ url: `/request/${id}`, method: 'get' }), {
       onSuccess: (data) => {
-        setViewport({ longitude: data.long, latitude: data.lat, zoom: 16 });
+        setViewport({ longitude: data.long, latitude: data.lat, zoom: 16, maxZoom: 16, minZoom: 13 });
       }
     }
   );
-
-  /*Comments query */
-  const {
-    isLoading: commentLoading,
-    isError: commentError,
-    data: commentData,
-    refetch,
-  } = useQuery('comments', () =>
-    useAxios({ url: `/comments/${id}`, method: 'get' }),
-  );
-
-  const refreshButton = useRef(null);
-
-  const [commentText, setText] = useState('');
-
-  const commentMutation = useMutation((newComment) =>
-    useAxios({
-      url: `/comments/${id}`,
-      method: 'post',
-      params: newComment,
-    }),
-  );
-
-  function postComment() {
-    commentMutation.mutate({
-      // avatar:
-      commentBody: commentText,
-      sender_id: currentUser.id,
-      request_id: id,
-    });
-    setText('');
-    setTimeout(() => {
-      refreshButton.current.click();
-      console.log('Refresh comments ', refreshButton.current);
-    }, 1500);
-  }
 
   function offerHelp() {
     mutation.mutate(
@@ -239,40 +203,11 @@ function RequestId({ id }) {
                 )}
             </div>
           )}
-          <div id="comments-container" className='flex mx-auto items-center border rounded-lg sm:px-6 sm:py-4 leading-relaxed'>
-            <form className='w-full max-w-xl bg-white rounded-lg px-4 pt-2'>
-              <div className='flex flex-col -mx-3 mb-6'>
-                <strong className='px-4 pt-3 pb-2 text-gray-800 text-lg'>
-                  Add a new comment
-                </strong>
-                <div className='flex flex-col'>
-                  <div className='w-full md:w-full px-3 mb-2 mt-2'>
-                    <textarea
-                      className='bg-gray-100 rounded border border-gray-400 leading-normal resize-none w-full h-20 py-2 px-3 font-medium placeholder-gray-700 focus:outline-none focus:bg-white'
-                      name='body'
-                      placeholder='Write something...'
-                      onChange={(event) => setText(event.target.value)}
-                      value={commentText}
-                      required></textarea>
-                  </div>
-                  {/* <!-- comment form --> */}
-                  <button
-                    type='submit'
-                    className='btn btn-secondary'
-                    value='Post Comment'
-                    disabled={commentText === ''}
-                    onClick={postComment}>
-                    Post Comment
-                  </button>
-                  <button
-                    className='-z-2'
-                    ref={refreshButton}
-                    onClick={refetch}></button>
-                </div>
-              </div>
-            </form>{' '}
-          </div>
-          <CommentList commentsData={commentData} />
+          {
+            data && (
+              <Comments id={data.id}/>
+            )
+          }
         </Container>
       </NavBar>
     </>
